Some fixes to release 0.5.0.

+++ src/sc_containers.h
@@ -260,7 +260,7 @@ sc_mempool_alloc (sc_mempool_t * mempool)
     ret = *(void **) sc_array_pop (freed);
   }
   else {
-    ret = obstack_alloc (&mempool->obstack, mempool->elem_size);
+    ret = obstack_alloc (&mempool->obstack, (int) mempool->elem_size);
   }
 
 #ifdef SC_DEBUG
